pyproject.toml
[project]
name = "organization-directory-api"
version = "1.0.0"
description = "REST API for Organization Directory"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.30.0",
    "environ-config>=24.1.0",
    "attrs>=24.2.0",
    "python-dotenv>=1.0.0",
    "geoalchemy2>=0.15.0",
    "shapely>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "ty>=0.0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RET",   # flake8-return
    "PTH",   # flake8-use-pathlib
    "ERA",   # eradicate (commented-out code)
    "C901",  # mccabe complexity
]
ignore = [
    "B008",   # do not perform function calls in argument defaults (FastAPI Depends)
    "UP046",  # Generic[T] subclass - Pydantic/SQLAlchemy don't fully support PEP 695
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["src/seed.py"]

[tool.coverage.report]
show_missing = true
fail_under = 70
